+++ python/GafferVDBUI/MeshToLevelSetUI.py
@@ -42,6 +42,34 @@ GafferUI.Metadata.registerNode(
 	'description',
 	"""Converts a mesh primitive to a level set VDB object.""",
 	plugs={
+		"filter" : [
+			"description",
+			"""
+			The filter used to choose the meshes to be converted. Source locations are
+			pruned from the output scene, unless they are reused as part of a destination location
+			(or a separate source scene is connected).
+			"""
+		],
+		"source" : [
+			"description",
+			"""
+			An optional alternate scene to provide the meshes to be converted. When connected :
+
+			- The `filter` chooses locations to be merged from the `source` scene rather than then `in` scene.
+			- Source locations are not pruned from the output scene.
+			"""
+		],
+		"destination" : [
+			"description",
+			"""
+			By default, meshes will be replaced with a level set in place, using the destination `${scene:path}`.
+			The destination can be modified to change where the outputs are placed. If multiple filtered locations
+			have the same destination, the meshes will be merged into one level set.
+
+			The destination location will be created if it doesn't exist already. If the name overlaps
+			with an existing location that isn't filtered, the name will get a suffix.
+			""",
+		],
 		'grid' : [
 			'description',
 			"""