+++ src/GafferScene/MergeCurves.cpp
@@ -63,7 +63,7 @@ MergeCurves::~MergeCurves()
 {
 }
 
-IECore::ConstObjectPtr MergeCurves::mergeObjects( const std::vector< std::pair< IECore::ConstObjectPtr, Imath::M44f > > &sources, const Gaffer::Context *context ) const
+IECore::ConstObjectPtr MergeCurves::computeMergedObject( const std::vector< std::pair< IECore::ConstObjectPtr, Imath::M44f > > &sources, const Gaffer::Context *context ) const
 {
 	std::vector< std::pair< const IECoreScene::Primitive *, Imath::M44f > > curves;