+++ src/GafferScene/MergePoints.cpp
@@ -63,7 +63,7 @@ MergePoints::~MergePoints()
 {
 }
 
-IECore::ConstObjectPtr MergePoints::mergeObjects( const std::vector< std::pair< IECore::ConstObjectPtr, Imath::M44f > > &sources, const Gaffer::Context *context ) const
+IECore::ConstObjectPtr MergePoints::computeMergedObject( const std::vector< std::pair< IECore::ConstObjectPtr, Imath::M44f > > &sources, const Gaffer::Context *context ) const
 {
 	std::vector< std::pair< const IECoreScene::Primitive *, Imath::M44f > > points;